Get 2nd matricule for GBM6125
